Refactor page components by consolidating logic into next js `page.tsx` files for improved structure and maintainability.
